requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
blis==1.2.1
blobfile==3.0.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
bs4==0.0.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cmudict==1.0.32
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
decorator==5.2.1
diffusers==0.33.1
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
easyocr==1.7.2
einops==0.8.1
emoji==1.2.0
encodec==0.1.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
g2pkk==0.1.2
grpcio==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-jamo==1.0.1
hangul-romanize==0.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.1.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kollocate==0.0.2
koparadigm==0.10.0
kss==6.0.4
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llama_models==0.2.0
llama_stack==0.2.3
llama_stack_client==0.2.3
llama_toolchain==0.0.35
llvmlite==0.43.0
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
murmurhash==1.0.12
networkx==2.8.8
ninja==1.11.1.4
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
openai==1.76.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==1.5.3
pecab==1.0.8
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pyaml==25.1.0
pyarrow==19.0.1
pybind11==2.9.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.54.0
pysbd==0.3.4
pytest==8.3.5
pytest-asyncio==0.26.0
python-bidi==0.6.6
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
sacremoses==0.1.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.11.4
sentence-transformers==4.1.0
sentencepiece==0.2.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tossi==0.3.1
tqdm==4.67.1
trainer==0.0.36
transformers==4.51.3
TTS==0.22.0
typeguard==4.4.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
Whoosh==2.7.4
wrapt==1.17.2
xlrd==1.2.0
yarl==1.20.0
zipp==3.21.0
